+++ koala/cogs/__init__.py
@@ -1,6 +1,7 @@
 from .announce import Announce
 from .base import BaseCog
 from .colour_role import ColourRole
+from .insights import Insights
 from .intro_cog import IntroCog
 from .react_for_role import ReactForRole
 from .text_filter import TextFilter