+++ koala/cogs/intro_cog/models.py
@@ -7,7 +7,7 @@ from koala.models import mapper_registry, DiscordSnowflake
 class GuildWelcomeMessages:
     __tablename__ = 'GuildWelcomeMessages'
     guild_id = Column(DiscordSnowflake, primary_key=True)
-    welcome_message = Column(VARCHAR(2000, collation="utf8mb4_general_ci"), nullable=True)
+    welcome_message = Column(VARCHAR(2000, collation="utf8mb4_unicode_520_ci"), nullable=True)
 
     def __repr__(self):
         return "<GuildWelcomeMessages(%s, %s)>" % \